Commented out the font size functionality
This wasn’t working when I removed it from the hovercar nested set up
because it was relying on variables that were in scope. I’d like to
change it so they rely on passed in arguments and then return values,
but I wasn’t able to get it working yet. WIP